review commit: * added a cache to SearchIndexingSignalEnrichmentFacade in order to only evaluate "patterns" once for a given namespace * removed copy&pasted unit tests in SearchIndexingSignalEnrichmentFacadeTest by adding another abstract test class AbstractCachingSignalEnrichmentFacadeTest * added a unit test testing selection of JsonFieldSelectors based on different namespaces * minor cleanup and formatting * enhanced documentation about how to configure the indexed namespaces via system properties
