+++ thingsearch/service/src/test/resources/namespace-search-index-test.conf
@@ -11,12 +11,19 @@ ditto {
     namespace-search-include-fields = [
       # The list of thing paths that are included in the search index.
       {
-        namespace-pattern = "org.eclipse",
-        search-include-fields = [ "attributes", "features/info" ]
+        namespace-pattern = "org.eclipse.test"
+        search-include-fields = [
+          "attributes",
+          "features/info/properties",
+          "features/info/other"
+        ]
       },
       {
-        namespace-pattern = "org.eclipse.test",
-        search-include-fields = [ "attributes", "features/info/properties/", "features/info/other" ]
+        namespace-pattern = "org.eclipse*"
+        search-include-fields = [
+          "attributes",
+          "features/info"
+        ]
       }
     ]
   }
